Major fixes for all warnings, added compat for windows

+++ include/utils/cifar10_data_loader.hpp
@@ -151,7 +151,7 @@ public:
    * Get a specific batch size (supports both pre-computed and on-demand
    * batches)
    */
-  bool get_batch(int batch_size, Tensor<T> &batch_data,
+  bool get_batch(size_t batch_size, Tensor<T> &batch_data,
                  Tensor<T> &batch_labels) override {
 
     if (batches_prepared_ && batch_size == this->batch_size_) {
@@ -162,19 +162,19 @@ public:
       return false;
     }
 
-    const int actual_batch_size = std::min(
-        batch_size, static_cast<int>(data_.size() - this->current_index_));
+    const size_t actual_batch_size = std::min(
+        batch_size, data_.size() - this->current_index_);
 
     batch_data = Tensor<T>(
-        static_cast<size_t>(actual_batch_size), cifar10_constants::NUM_CHANNELS,
+        actual_batch_size, cifar10_constants::NUM_CHANNELS,
         cifar10_constants::IMAGE_HEIGHT, cifar10_constants::IMAGE_WIDTH);
 
-    batch_labels = Tensor<T>(static_cast<size_t>(actual_batch_size),
+    batch_labels = Tensor<T>(actual_batch_size,
                              cifar10_constants::NUM_CLASSES, 1, 1);
     batch_labels.fill(static_cast<T>(0.0));
 
 #pragma omp parallel for if (actual_batch_size > 16)
-    for (int i = 0; i < actual_batch_size; ++i) {
+    for (size_t i = 0; i < actual_batch_size; ++i) {
       const auto &image_data = data_[this->current_index_ + i];
 
       for (int c = 0; c < static_cast<int>(cifar10_constants::NUM_CHANNELS);
@@ -284,7 +284,7 @@ public:
   /**
    * Pre-compute all batches for efficient training
    */
-  void prepare_batches(int batch_size) override {
+  void prepare_batches(size_t batch_size) override {
     if (data_.empty()) {
       std::cerr << "Warning: No data loaded, cannot prepare batches!"
                 << std::endl;
@@ -308,28 +308,24 @@ public:
     for (size_t batch_idx = 0; batch_idx < num_batches; ++batch_idx) {
       const size_t start_idx = batch_idx * batch_size;
       const size_t end_idx = std::min(start_idx + batch_size, num_samples);
-      const int actual_batch_size = static_cast<int>(end_idx - start_idx);
+      const size_t actual_batch_size = end_idx - start_idx;
 
-      Tensor<T> batch_data(static_cast<size_t>(actual_batch_size),
-                           cifar10_constants::NUM_CHANNELS,
+      Tensor<T> batch_data(actual_batch_size, cifar10_constants::NUM_CHANNELS,
                            cifar10_constants::IMAGE_HEIGHT,
                            cifar10_constants::IMAGE_WIDTH);
 
-      Tensor<T> batch_labels(static_cast<size_t>(actual_batch_size),
+      Tensor<T> batch_labels(actual_batch_size,
                              cifar10_constants::NUM_CLASSES, 1, 1);
       batch_labels.fill(static_cast<T>(0.0));
 
 #pragma omp parallel for if (actual_batch_size > 16)
-      for (int i = 0; i < actual_batch_size; ++i) {
+      for (size_t i = 0; i < actual_batch_size; ++i) {
         const size_t sample_idx = start_idx + i;
         const auto &image_data = data_[sample_idx];
 
-        for (int c = 0; c < static_cast<int>(cifar10_constants::NUM_CHANNELS);
-             ++c) {
-          for (int h = 0; h < static_cast<int>(cifar10_constants::IMAGE_HEIGHT);
-               ++h) {
-            for (int w = 0;
-                 w < static_cast<int>(cifar10_constants::IMAGE_WIDTH); ++w) {
+        for (size_t c = 0; c < cifar10_constants::NUM_CHANNELS; ++c) {
+          for (size_t h = 0; h < cifar10_constants::IMAGE_HEIGHT; ++h) {
+            for (size_t w = 0; w < cifar10_constants::IMAGE_WIDTH; ++w) {
               size_t pixel_idx = c * cifar10_constants::IMAGE_HEIGHT *
                                      cifar10_constants::IMAGE_WIDTH +
                                  h * cifar10_constants::IMAGE_WIDTH + w;
